secrets init fix
