+++ command/secrets/init/params_test.go
@@ -1,4 +1,4 @@
-package polybftsecrets
+package init
 
 import (
 	"encoding/hex"
