+++ command/secrets/init/result.go
@@ -4,15 +4,32 @@ import (
 	"bytes"
 	"fmt"
 
+	"github.com/0xPolygon/polygon-edge/command"
+
 	"github.com/0xPolygon/polygon-edge/command/helper"
 	"github.com/0xPolygon/polygon-edge/types"
 )
 
+type Results []command.CommandResult
+
+func (r Results) GetOutput() string {
+	var buffer bytes.Buffer
+
+	for _, result := range r {
+		buffer.WriteString(result.GetOutput())
+	}
+
+	return buffer.String()
+}
+
 type SecretsInitResult struct {
-	Address   types.Address `json:"address"`
-	BLSPubkey string        `json:"bls_pubkey"`
-	NodeID    string        `json:"node_id"`
-	Insecure  bool          `json:"insecure"`
+	Address       types.Address `json:"address"`
+	BLSPubkey     string        `json:"bls_pubkey"`
+	NodeID        string        `json:"node_id"`
+	PrivateKey    string        `json:"private_key"`
+	BLSPrivateKey string        `json:"bls_private_key"`
+	Insecure      bool          `json:"insecure"`
+	Generated     string        `json:"generated"`
 }
 
 func (r *SecretsInitResult) GetOutput() string {
@@ -25,6 +42,20 @@ func (r *SecretsInitResult) GetOutput() string {
 		fmt.Sprintf("Public key (address)|%s", r.Address.String()),
 	)
 
+	if r.PrivateKey != "" {
+		vals = append(
+			vals,
+			fmt.Sprintf("Private key|%s", r.PrivateKey),
+		)
+	}
+
+	if r.BLSPrivateKey != "" {
+		vals = append(
+			vals,
+			fmt.Sprintf("BLS Private key|%s", r.BLSPrivateKey),
+		)
+	}
+
 	if r.BLSPubkey != "" {
 		vals = append(
 			vals,
@@ -36,6 +67,12 @@ func (r *SecretsInitResult) GetOutput() string {
 
 	if r.Insecure {
 		buffer.WriteString("\n[WARNING: INSECURE LOCAL SECRETS - SHOULD NOT BE RUN IN PRODUCTION]\n")
+	}
+
+	if r.Generated != "" {
+		buffer.WriteString("\n[SECRETS GENERATED]\n")
+		buffer.WriteString(r.Generated)
+		buffer.WriteString("\n")
 	}
 
 	buffer.WriteString("\n[SECRETS INIT]\n")
