+++ command/secrets/init/utils.go
@@ -1,4 +1,4 @@
-package polybftsecrets
+package init
 
 import (
 	"errors"
