+++ command/server/params.go
@@ -47,12 +47,6 @@ const (
 	metricsIntervalFlag = "metrics-interval"
 )
 
-// Flags that are deprecated, but need to be preserved for
-// backwards compatibility with existing scripts
-const (
-	ibftBaseTimeoutFlagLEGACY = "ibft-base-timeout"
-)
-
 const (
 	unsetPeersValue = -1
 )
