+++ scripts/cluster
@@ -32,7 +32,6 @@ fi
 function showhelp() {
   echo "Usage: cluster {consensus} [{command}] [{flags}]"
   echo "Consensus:"
-  echo "  ibft            Start Supernets test environment locally with ibft consensus"
   echo "  polybft         Start Supernets test environment locally with polybft consensus"
   echo "Commands:"
   echo "  stop            Stop the running environment"
@@ -47,26 +46,14 @@ function showhelp() {
   echo "  cluster polybft stop -- Stop the running environment"
 }
 
-function initIbftConsensus() {
-  echo "Running with ibft consensus"
-  ./polygon-edge secrets init --insecure --data-dir test-chain- --num 4
-
-  node1_id=$(./polygon-edge secrets output --data-dir test-chain-1 | grep Node | head -n 1 | awk -F ' ' '{print $4}')
-  node2_id=$(./polygon-edge secrets output --data-dir test-chain-2 | grep Node | head -n 1 | awk -F ' ' '{print $4}')
-
-  genesis_params="--consensus ibft --validators-prefix test-chain- \
-    --bootnode /ip4/127.0.0.1/tcp/30301/p2p/$node1_id \
-    --bootnode /ip4/127.0.0.1/tcp/30302/p2p/$node2_id"
-}
-
 function initPolybftConsensus() {
   echo "Running with polybft consensus"
   genesis_params="--consensus polybft"
 
-  address1=$(./polygon-edge polybft-secrets --insecure --data-dir test-chain-1 | grep Public | head -n 1 | awk -F ' ' '{print $5}')
-  address2=$(./polygon-edge polybft-secrets --insecure --data-dir test-chain-2 | grep Public | head -n 1 | awk -F ' ' '{print $5}')
-  address3=$(./polygon-edge polybft-secrets --insecure --data-dir test-chain-3 | grep Public | head -n 1 | awk -F ' ' '{print $5}')
-  address4=$(./polygon-edge polybft-secrets --insecure --data-dir test-chain-4 | grep Public | head -n 1 | awk -F ' ' '{print $5}')
+  address1=$(./polygon-edge secrets init --insecure --data-dir test-chain-1 | grep Public | head -n 1 | awk -F ' ' '{print $5}')
+  address2=$(./polygon-edge secrets init --insecure --data-dir test-chain-2 | grep Public | head -n 1 | awk -F ' ' '{print $5}')
+  address3=$(./polygon-edge secrets init --insecure --data-dir test-chain-3 | grep Public | head -n 1 | awk -F ' ' '{print $5}')
+  address4=$(./polygon-edge secrets init --insecure --data-dir test-chain-4 | grep Public | head -n 1 | awk -F ' ' '{print $5}')
 }
 
 function createGenesis() {
@@ -254,15 +241,8 @@ case "$2" in
 # cluster {consensus}
 *)
   echo "Running $1 environment from local binary..."
-  # Initialize ibft or polybft consensus
-  if [ "$1" == "ibft" ]; then
-    # Initialize ibft consensus
-    initIbftConsensus
-    # Create genesis file and start the server from binary
-    createGenesis
-    startNodes $1 $2
-    exit 0
-  elif [ "$1" == "polybft" ]; then
+  # Initialize polybft consensus
+  if [ "$1" == "polybft" ]; then
     # Initialize polybft consensus
     initPolybftConsensus
     # Create genesis file and start the server from binary
@@ -271,7 +251,7 @@ case "$2" in
     startNodes $1 $2
     exit 0
   else
-    echo "Unsupported consensus mode. Supported modes are: ibft and polybft."
+    echo "Unsupported consensus mode. Supported modes are: polybft."
     showhelp
     exit 1
   fi
